python/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
pandas==2.3.2
numpy==2.3.2
scikit-learn==1.7.1
MetaTrader5==5.0.47; platform_system == "Windows"
loguru==0.7.3
httpx==0.28.1
streamlit==1.48.1
plotly==6.3.0
requests==2.32.4
joblib==1.5.2
